feat: add tests base class
